+++ hari_client/models/models.py
@@ -14,7 +14,13 @@ typeT = typing.TypeVar("typeT", bound=typing.Any)
 
 
 class BaseModel(pydantic.BaseModel):
-    model_config = pydantic.ConfigDict(extra="allow")
+    model_config = pydantic.ConfigDict(
+        extra="allow",
+        json_encoders={
+            pydantic.BaseModel: lambda v: v.model_dump(),  # Serialize all nested Pydantic models using .model_dump()
+            uuid.UUID: lambda v: str(v),
+        },
+    )
 
 
 class VideoParameters(str, enum.Enum):
@@ -143,7 +149,7 @@ class CuboidCenterPoint(BaseModel):
     """A 3D cuboid defined by its center point position, heading as quaternion and its
     dimensions along each axis."""
 
-    type: str = pydantic.Field(title="Type")
+    type: str = "cuboid_center_point"
     position: Point3DTuple = pydantic.Field()
     heading: QuaternionTuple = pydantic.Field()
     dimensions: Point3DTuple = pydantic.Field()
@@ -258,6 +264,33 @@ class VisualisationType(str, enum.Enum):
     IMAGETRANSFORMATION = "ImageTransformation"
     VIDEO = "Video"
     RENDERED = "Rendered"
+
+
+class MLAnnotationModelStatus(str, enum.Enum):
+    CREATED = "created"
+    TRAINING = "training"
+    TRAINING_FAILED = "training_failed"
+    TRAINING_DONE = "training_done"
+
+
+class AIAnnotationRunStatus(str, enum.Enum):
+    CREATED = "created"
+    ANNOTATING = "annotating"
+    CREATION_FAILED = "creation_failed"
+    AI_ANNOTATION_FAILED = "ai_annotation_failed"
+    DONE = "done"
+
+
+class AINTLearningDataStatus(str, enum.Enum):
+    SUBMITTED = "submitted"
+    VALIDATING = "validating"
+    VALIDATION_FAILED = "validation_failed"
+    CREATED = "created"
+
+
+class AnnotationAnswer(BaseModel):
+    annotation_id: str
+    value: typeT
 
 
 class ExternalMediaSourceCredentialsType(str, enum.Enum):
@@ -396,7 +429,7 @@ class CameraIntrinsics(BaseModel):
 
 class PointCloudMetadata(BaseModel):
     sensor_id: str = pydantic.Field(title="Sensor Id")
-    lidar_sensor_pose: dict[str, typing.Any] = pydantic.Field(title="Lidar Sensor Pose")
+    lidar_sensor_pose: dict[str, Pose3D] = pydantic.Field(title="Lidar Sensor Pose")
 
 
 class ImageMetadata(BaseModel):
@@ -406,6 +439,28 @@ class ImageMetadata(BaseModel):
         default=None, title="CameraIntrinsics"
     )
     camera_extrinsics: Pose3D | None = pydantic.Field(default=None, title="Pose3D")
+
+
+class SceneCreate(BaseModel):
+    back_reference: str
+    frames: list[Scene.Frame]
+
+
+class Frame(BaseModel):
+    index: int = pydantic.Field(title="Index")
+
+
+class Scene(BaseModel):
+    id: str = pydantic.Field(title="Id")
+    dataset_id: str = pydantic.Field(title="Dataset Id")
+    tags: list[str] | None = pydantic.Field(default=None, title="Tags")
+    timestamp: str = pydantic.Field(
+        default=None,
+        title="Timestamp",
+    )
+    archived: bool | None = pydantic.Field(default=None, title="Archived")
+    back_reference: str | None = pydantic.Field(default=None, title="Back Reference")
+    frames: list[Frame] = pydantic.Field(default=[], title="Frames")
 
 
 class TransformationParameters(BaseModel):
@@ -553,7 +608,9 @@ class Media(BaseModel):
     archived: bool | None = pydantic.Field(default=False, title="Archived")
     back_reference: str = pydantic.Field(title="Back Reference")
     subset_ids: list = pydantic.Field(default=[], title="Subset Ids")
-    attributes: list = pydantic.Field(default=[], title="Attributes")
+    attributes: list[AttributeValueResponse] = pydantic.Field(
+        default=[], title="Attributes"
+    )
     thumbnails: dict[str, typing.Any] = pydantic.Field(default={}, title="Thumbnails")
     visualisations: list[VisualisationUnion] | None = pydantic.Field(
         default=None, title="Visualisations"
@@ -563,9 +620,9 @@ class Media(BaseModel):
         default=None, title="Realworldobject Id"
     )
     type: str = pydantic.Field(default="Media", title="Type")
-    media_url: str = pydantic.Field(title="Media Url")
+    media_url: str | None = pydantic.Field(default=None, title="Media Url")
     file_key: str | None = pydantic.Field(default=None, title="File Key")
-    pii_media_url: str = pydantic.Field(title="Pii Media Url")
+    pii_media_url: str | None = pydantic.Field(default=None, title="Pii Media Url")
     name: str = pydantic.Field(title="Name")
     metadata: ImageMetadata | PointCloudMetadata | None = pydantic.Field(
         default=None, title="ImageMetadata"
@@ -586,7 +643,9 @@ class MediaResponse(BaseModel):
     archived: bool | None = pydantic.Field(default=None, title="Archived")
     back_reference: str | None = pydantic.Field(default=None, title="Back Reference")
     subset_ids: list | None = pydantic.Field(default=None, title="Subset Ids")
-    attributes: list | None = pydantic.Field(default=None, title="Attributes")
+    attributes: list[AttributeValueResponse] | None = pydantic.Field(
+        default=None, title="Attributes"
+    )
     thumbnails: dict[str, typing.Any] | None = pydantic.Field(
         default=None, title="Thumbnails"
     )
@@ -710,7 +769,9 @@ class MediaObject(BaseModel):
     archived: bool | None = pydantic.Field(default=False, title="Archived")
     back_reference: str = pydantic.Field(title="Back Reference")
     subset_ids: list = pydantic.Field(default=[], title="Subset Ids")
-    attributes: list = pydantic.Field(default=[], title="Attributes")
+    attributes: list[AttributeValueResponse] = pydantic.Field(
+        default=[], title="Attributes"
+    )
     thumbnails: dict[str, typing.Any] = pydantic.Field(default={}, title="Thumbnails")
     visualisations: list[VisualisationUnion] | None = pydantic.Field(
         default=None, title="Visualisations"
@@ -746,7 +807,9 @@ class MediaObjectResponse(BaseModel):
     archived: bool | None = pydantic.Field(default=None, title="Archived")
     back_reference: str | None = pydantic.Field(default=None, title="Back Reference")
     subset_ids: list | None = pydantic.Field(default=None, title="Subset Ids")
-    attributes: list | None = pydantic.Field(default=None, title="Attributes")
+    attributes: list[AttributeValueResponse] | None = pydantic.Field(
+        default=None, title="Attributes"
+    )
     thumbnails: dict[str, typing.Any] | None = pydantic.Field(
         default=None, title="Thumbnails"
     )
@@ -773,6 +836,103 @@ class MediaObjectResponse(BaseModel):
     instance_id: str | None = pydantic.Field(default=None, title="Instance Id")
     media_object_type: MediaObjectType | None = pydantic.Field(
         default=None, title="Media Object Type"
+    )
+
+
+class TrainingAttribute(BaseModel):
+    dataset_id: uuid.UUID = pydantic.Field(default=None, title="Dataset Id")
+    attribute_id: str = pydantic.Field(default=None, title="Attribute Id")
+    query: QueryList | None = pydantic.Field(default_factory=list, title="Query")
+
+
+class AINTLearningData(BaseModel):
+    id: uuid.UUID = pydantic.Field(title="Id")
+    name: str = pydantic.Field(title="Name")
+    created_at: datetime.datetime | None = pydantic.Field(
+        title="Created At", default=None
+    )
+    updated_at: datetime.datetime | None = pydantic.Field(
+        title="Updated At", default=None
+    )
+    archived_at: datetime.datetime | None = pydantic.Field(
+        title="Archived At", default=None
+    )
+    owner: str | None = pydantic.Field(default=None, title="Owner")
+    user_group: str | None = pydantic.Field(default=None, title="User Group")
+    training_attributes: list[TrainingAttribute] = pydantic.Field(
+        default_factory=list, title="Training Attributes"
+    )
+    question: str = pydantic.Field(title="Question")
+    possible_answers: list[str] = pydantic.Field(
+        default_factory=list, title="Possible Answers"
+    )
+    repeats: int = pydantic.Field(title="Repeats")
+    subset_id: uuid.UUID = pydantic.Field(title="Subset Id")
+    status: AINTLearningDataStatus = pydantic.Field(title="Status")
+
+
+class MlAnnotationModel(BaseModel):
+    id: uuid.UUID = pydantic.Field(title="Id")
+    created_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Created At"
+    )
+    updated_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Updated At"
+    )
+    archived_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Archived At"
+    )
+    owner: str | None = pydantic.Field(default=None, title="Owner")
+    user_group: str | None = pydantic.Field(default=None, title="User Group")
+    status: MLAnnotationModelStatus = pydantic.Field(title="Status")
+    dataset_id: str = pydantic.Field(title="Dataset Id")
+    reference_set_annotation_run_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Reference Set Annotation Run Id"
+    )
+    name: str = pydantic.Field(title="Name")
+    training_subset_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Training Subset Id"
+    )
+    validation_subset_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Validation Subset Id"
+    )
+    test_subset_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Test Subset Id"
+    )
+    automation_correctness_curve: dict | None = pydantic.Field(
+        default_factory=dict, title="Automation Correctness Curve"
+    )
+    model_weight_location: str | None = pydantic.Field(
+        default=None, title="Model Weight Location"
+    )
+    aint_learning_data_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="AINT learning data Id"
+    )
+
+
+class AIAnnotationRun(BaseModel):
+    id: uuid.UUID = pydantic.Field(title="Id")
+    created_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Created At"
+    )
+    updated_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Updated At"
+    )
+    archived_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Archived At"
+    )
+    owner: str | None = pydantic.Field(default=None, title="Owner")
+    user_group: str | None = pydantic.Field(default=None, title="User Group")
+    name: str = pydantic.Field(title="Name")
+    status: AIAnnotationRunStatus = pydantic.Field(title="Status")
+    dataset_id: uuid.UUID = pydantic.Field(title="Dataset Id")
+    subset_id: uuid.UUID = pydantic.Field(title="Subset Id")
+    ml_annotation_model_id: uuid.UUID = pydantic.Field(title="ML Annotation Model Id")
+    attribute_metadata_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Attribute Metadata Id"
+    )
+    completed_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Completed At"
     )
 
 
@@ -1040,6 +1200,8 @@ class AttributeCreate(BaseModel):
     cumulated_frequency: typing.Any | None = None
     frequency: dict[str, int] | None = None
     question: str | None = None
+    ml_predictions: dict[str, float] | None = None
+    ml_probability_distributions: dict[str, float] | None = None
     repeats: int | None = None
     possible_values: list[str] | None = None
 
@@ -1054,43 +1216,10 @@ class AttributeCreate(BaseModel):
         return values
 
 
-class Attribute(BaseModel):
-    id: str = pydantic.Field(title="ID")
-    name: str
-    question: str
-    annotatable_id: str
-    annotatable_type: DataBaseObjectType
-    subset_ids: set[str] = set()
-    attribute_type: AttributeType | None = None
-    attribute_group: AttributeGroup
-    value: typeT
-    min: typeT | None = None
-    max: typeT | None = None
-    sum: typeT | None = None
-    cant_solves: int | None = None
-    solvability: float | None = None
-    aggregate: typeT | None = None
-    modal: typeT | None = None
-    credibility: float | None = None
-    convergence: float | None = None
-    ambiguity: float | None = None
-    median: typeT | None = None
-    variance: float | None = None
-    standard_deviation: float | None = None
-    range: typing.Any | None = None
-    average_absolute_deviation: float | None = None
-    cumulated_frequency: typing.Any | None = None
-    frequency: dict[str, int] | None = None
-    question: str | None = None
-    ml_predictions: dict[str, float] | None = None
-    ml_probability_distributions: dict[str, float] | None = None
-    cant_solve_ratio: float | None = None
-    repeats: int | None = None
-    possible_values: list[str] | None = None
-
-
 class AttributeResponse(BaseModel):
+    # AttributeValue + AttributeMetadata = Attribute
     id: str | None = pydantic.Field(default=None, title="Id")
+    tags: list | None = pydantic.Field(default=None, title="Tags")
     dataset_id: str | None = pydantic.Field(default=None, title="Dataset Id")
     timestamp: str | None = pydantic.Field(default=None, title="Timestamp")
     archived: bool | None = pydantic.Field(default=None, title="Archived")
@@ -1108,12 +1237,22 @@ class AttributeResponse(BaseModel):
     attribute_group: AttributeGroup | None = pydantic.Field(
         default=None, title="Attribute Group"
     )
+    pipeline_project: dict | None = pydantic.Field(
+        default=None, title="Pipeline Project"
+    )
+    annotation_run_node_id: str | None = pydantic.Field(
+        default=None, title="Annotation Run Node ID"
+    )
+    annotation_run_id: str | None = pydantic.Field(
+        default=None, title="Annotation Run ID"
+    )
     value: typeT | None = pydantic.Field(default=None, title="Value")
     min: typeT | None = pydantic.Field(default=None, title="Min")
     max: typeT | None = pydantic.Field(default=None, title="Max")
     sum: typeT | None = pydantic.Field(default=None, title="Sum")
     cant_solves: int | None = pydantic.Field(default=None, title="Cant Solves")
     solvability: float | None = pydantic.Field(default=None, title="Solvability")
+    confidence: float | None = pydantic.Field(default=None, title="Confidence")
     aggregate: typeT | None = pydantic.Field(default=None, title="Aggregate")
     modal: typeT | None = pydantic.Field(default=None, title="Modal")
     credibility: float | None = pydantic.Field(default=None, title="Credibility")
@@ -1158,17 +1297,101 @@ class AttributeResponse(BaseModel):
         title="Possible Values",
         description="Possible values for this attribute",
     )
+    annotations: list[AnnotationAnswer] | None = pydantic.Field(
+        default=None,
+        title="Annotations",
+        description="Annotations for this attribute",
+    )
+    visualisation_id: str | None = pydantic.Field(
+        default=None,
+        title="Visualisation ID",
+        description="Visualisation ID for this attribute",
+    )
+    visualisation_config_id: str | None = pydantic.Field(
+        default=None,
+        title="Visualisation Config ID",
+        description="Visualisation Config ID for this attribute",
+    )
+
+
+class AttributeValueResponse(BaseModel):
+    # AttributeValue + AttributeMetadata = Attribute
+    id: str | None = pydantic.Field(default=None, title="Id")
+    dataset_id: str | None = pydantic.Field(default=None, title="Dataset Id")
+    tags: list | None = pydantic.Field(default=None, title="Tags")
+    timestamp: str | None = pydantic.Field(default=None, title="Timestamp")
+    archived: bool | None = pydantic.Field(default=None, title="Archived")
+    annotatable_id: str | None = pydantic.Field(default=None, title="Annotatable ID")
+    metadata_id: str | None = pydantic.Field(default=None, title="Metadata Id")
+    annotatable_type: DataBaseObjectType | None = pydantic.Field(
+        default=None, title="Annotatable Type"
+    )
+    value: typeT | None = pydantic.Field(default=None, title="Value")
+    min: typeT | None = pydantic.Field(default=None, title="Min")
+    max: typeT | None = pydantic.Field(default=None, title="Max")
+    sum: typeT | None = pydantic.Field(default=None, title="Sum")
+    cant_solves: int | None = pydantic.Field(default=None, title="Cant Solves")
+    cant_solve_ratio: float | None = pydantic.Field(
+        default=None, title="Can't Solve Ratio"
+    )
+    solvability: float | None = pydantic.Field(default=None, title="Solvability")
+    aggregate: typeT | None = pydantic.Field(default=None, title="Aggregate")
+    modal: typeT | None = pydantic.Field(default=None, title="Modal")
+    credibility: float | None = pydantic.Field(default=None, title="Credibility")
+    convergence: float | None = pydantic.Field(default=None, title="Convergence")
+    confidence: float | None = pydantic.Field(default=None, title="Confidence")
+    ambiguity: float | None = pydantic.Field(default=None, title="Ambiguity")
+    median: typeT | None = pydantic.Field(default=None, title="Median")
+    variance: float | None = pydantic.Field(default=None, title="Variance")
+    standard_deviation: float | None = pydantic.Field(
+        default=None, title="Standard Deviation"
+    )
+    range: typing.Any | None = pydantic.Field(default=None, title="Range")
+    average_absolute_deviation: float | None = pydantic.Field(
+        default=None, title="Average Absolute Deviation"
+    )
+    cumulated_frequency: typing.Any | None = pydantic.Field(
+        default=None, title="Cumulated Frequency"
+    )
+    frequency: dict[str, int] | None = pydantic.Field(default=None, title="Frequency")
+    ml_predictions: dict[str, float] | None = pydantic.Field(
+        default=None,
+        title="ML Predictions",
+        description="These are the parameters of the posterior Dirichlet distribution",
+    )
+    ml_probability_distributions: dict[str, float] | None = pydantic.Field(
+        default=None,
+        title="ML Probability Distributions",
+        description="A point estimate for the probability associated with each category,"
+        " obtained from the full Dirichlet distribution predicted by the model.",
+    )
+    annotations: list[AnnotationAnswer] | None = pydantic.Field(
+        default=None,
+        title="Annotations",
+        description="Annotations for this attribute",
+    )
+    visualisation_id: str | None = pydantic.Field(
+        default=None,
+        title="Visualisation ID",
+        description="Visualisation ID for this attribute",
+    )
+    visualisation_config_id: str | None = pydantic.Field(
+        default=None,
+        title="Visualisation Config ID",
+        description="Visualisation Config ID for this attribute",
+    )
 
 
 class AttributeMetadataResponse(BaseModel):
+    # AttributeValue + AttributeMetadata = Attribute
     id: str | None = pydantic.Field(default=None, title="Id")
     dataset_id: str | None = pydantic.Field(default=None, title="Dataset Id")
-    tags: set[str] | None = pydantic.Field(default=set(), title="Tags")
+    tags: set[str] | None = pydantic.Field(default=None, title="Tags")
     timestamp: str | None = pydantic.Field(default=None, title="Timestamp")
     archived: bool | None = pydantic.Field(default=None, title="Archived")
     name: str | None = pydantic.Field(default=None, title="Name")
     question: str | None = pydantic.Field(default=None, title="Question")
-    subset_ids: set[str] = pydantic.Field(default=set(), title="Subset Ids")
+    subset_ids: set[str] | None = pydantic.Field(default=None, title="Subset Ids")
     attribute_type: AttributeType | None = pydantic.Field(
         default=None, title="Attribute Type"
     )
@@ -1187,6 +1410,269 @@ class AttributeMetadataResponse(BaseModel):
     pipeline_project: dict | None = pydantic.Field(
         default=None, title="Pipeline Project"
     )
+    possible_values: list[str] | None = pydantic.Field(
+        default=None,
+        title="Possible Values",
+        description="Possible values for this attribute",
+    )
+    repeats: int | None = pydantic.Field(
+        default=None,
+        title="Repeats",
+        description="Number of repeats for this attribute",
+    )
+
+
+class PipelineConfig(BaseModel):
+    limit_tasks: int | None = None
+    shuffle_tasks: bool = False
+    add_tasks_n_times: int = 1
+
+
+class Pipeline(BaseModel):
+    id: uuid.UUID = pydantic.Field(title="Id")
+    name: str = pydantic.Field(title="Name")
+    config: PipelineConfig | None = pydantic.Field(title="Config")
+    created_at: datetime.datetime = pydantic.Field(
+        default=None,
+        title="Created At",
+    )
+    owner: str | None = pydantic.Field(default=None, title="Owner")
+    updated_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Updated At"
+    )
+    deleted_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Deleted At"
+    )
+    user_group: str | None = pydantic.Field(default=None, title="User Group")
+    is_multinode: bool = False
+
+
+class PipelineNodeTypes(str, enum.Enum):
+    LEAF_NODE_LEGACY = "leaf_node"
+    DEFAULT_NODE = "node"
+    ROOT_NODE = "root_node"
+
+
+class VisualisationType(str, enum.Enum):
+    DEFAULT = "Default"  # full image
+    CROP = "Crop"
+    TILE = "Tile"
+    IMAGE_TRANSFORMATION = "ImageTransformation"
+    VIDEO = "Video"
+    RENDERED = "Rendered"
+
+
+class PipelineNodeConfig(pydantic.BaseModel):
+    node_type: PipelineNodeTypes = PipelineNodeTypes.DEFAULT_NODE
+    gui_settings: dict | None = None
+    media_type: str | None = None
+    annotatable_type: str | None = None
+    vendor_type: str | None = None
+    min_repeats: int | None = None
+    max_repeats: int | None = None
+    wp_tasks: int | None = None
+    wp_timeout: int | None = None
+    wp_rd_tasks: int | None = 0
+    presign_s3_urls: bool = True
+    should_workpackage_include_task_outputs: typing.Literal[
+        "parallel", "sequential"
+    ] = "parallel"
+    visualisation_type: VisualisationType | None = None
+    gui_type: str | None = None
+    possible_answers: list[dict[str, str]] | None = None
+    cant_solve_options: list[dict[str, str]] | None = None
+    split_map: dict[str, list[str]] = {}
+    convergence_threshold: float | None = pydantic.Field(None, ge=0, le=1)
+    auto_annotate: bool = False
+
+
+class PipelineNode(pydantic.BaseModel):
+    created_at: datetime.datetime
+    updated_at: datetime.datetime | None = None
+    archived_at: datetime.datetime | None = None
+    owner: uuid.UUID | None = None
+    user_group: str | None = None
+    name: str | None = None
+    config: PipelineNodeConfig | None = None
+    pipeline_id: uuid.UUID | None = None
+    id: uuid.UUID = pydantic.Field(title="Id")
+
+
+class PipelineWithNodes(Pipeline):
+    nodes: list[PipelineNode] | None = pydantic.Field(default=None, title="nodes")
+    root_node: PipelineNode | None = pydantic.Field(default=None, title="root_node")
+
+
+class MLAnnotationModelStatus(str, enum.Enum):
+    CREATED = "created"
+    TRAINING = "training"
+    TRAINING_FAILED = "training_failed"
+    TRAINING_DONE = "training_done"
+
+
+class MLAnnotationModel(BaseModel):
+    created_at: datetime.datetime = pydantic.Field(title="Created At")
+    updated_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Updated At"
+    )
+    archived_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Archived At"
+    )
+    owner: uuid.UUID | None = pydantic.Field(default=None, title="Owner")
+    user_group: str | None = pydantic.Field(default=None, title="User Group")
+    status: MLAnnotationModelStatus = pydantic.Field(title="Status")
+    dataset_id: uuid.UUID = pydantic.Field(title="Dataset Id")
+    reference_set_annotation_run_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Reference Set Annotation Run Id"
+    )
+    name: str = pydantic.Field(default="", title="Name")
+    training_subset_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Training Subset Id"
+    )
+    validation_subset_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Validation Subset Id"
+    )
+    test_subset_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Test Subset Id"
+    )
+    automation_correctness_curve: dict | None = pydantic.Field(
+        default=None, title="Automation Correctness Curve"
+    )
+    model_weight_location: str | None = pydantic.Field(
+        default=None, title="Model Weight Location"
+    )
+    training_set_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Training Set Id"
+    )
+    id: uuid.UUID = pydantic.Field(title="Id")
+
+    model_config = pydantic.ConfigDict(extra="ignore")
+
+
+class AnnotationRunStatus(str, enum.Enum):
+    CREATED = "created"
+    STARTED = "started"
+    ANNOTATION_DONE = "annotation_done"
+    POST_PROCESSING = "post_processing"
+    POST_PROCESSING_FAILED = "post_processing_failed"
+    AI_ANNOTATION_FAILED = "ai_annotation_failed"
+    AUTO_ANNOTATION_FAILED = "auto_annotation_failed"
+    CREATION_FAILED = "creation_failed"
+    PARTIALLY_DONE = "partially_done"
+    DONE = "done"
+
+
+class AnnotationRunType(str, enum.Enum):
+    MANUAL = "manual"
+    AI = "ai"
+
+
+class AnnotationRunNodeStatus(str, enum.Enum):
+    """
+    Annotation run statuses are duplicated in the qm-post-processing
+    Please make sure to keep them all in sync
+    """
+
+    CREATED = "created"
+    STARTED = "started"
+    POST_PROCESSING = "post_processing"
+    POST_PROCESSING_FAILED = "post_processing_failed"
+    AI_ANNOTATION_FAILED = "ai_annotation_failed"
+    AUTO_ANNOTATION_FAILED = "auto_annotation_failed"
+    CREATION_FAILED = "creation_failed"
+    ANNOTATION_DONE = "annotation_done"
+    POST_PROCESSING_DONE = "post_processing_done"
+
+
+class AnnotationRunNodeCreate(BaseModel):
+    id: uuid.UUID = pydantic.Field(default_factory=uuid.uuid4)
+    pipeline_node_id: uuid.UUID
+    status: AnnotationRunNodeStatus = AnnotationRunNodeStatus.CREATED
+
+
+class AnnotationRunNodeConfig(BaseModel):
+    reference_data_id: str | None = None
+    visualisation_config_id: str | None = None
+    override_min_repeats: int | None = None
+    override_max_repeats: int | None = None
+    override_wp_tasks: int | None = None
+    override_wp_timeout: int | None = None
+    initial_attribute_id: str | None = None
+
+
+class AnnotationRunNode(BaseModel):
+    id: uuid.UUID
+    created_at: datetime.datetime
+    updated_at: datetime.datetime | None
+    archived_at: datetime.datetime | None
+    owner: uuid.UUID
+    user_group: str | None
+    name: str | None
+    config: AnnotationRunNodeConfig | None = None
+    status: AnnotationRunNodeStatus | None
+    annotation_run_id: uuid.UUID
+    pipeline_node_revision_id: uuid.UUID
+
+
+class AnnotationRunConfig(pydantic.BaseModel):
+    color_map: dict[str, str] | None = None
+
+
+class AnnotationRunCreate(pydantic.BaseModel):
+    name: str
+    config: AnnotationRunConfig | None = pydantic.Field(
+        default=AnnotationRunConfig(), title="Config"
+    )
+    dataset_id: str
+    subset_id: str
+    pipeline_id: str
+    user_group: str | None = None
+    nodes: list[AnnotationRunNodeCreate] | None = pydantic.Field(
+        default=None, title="Nodes"
+    )
+
+
+class AnnotationRun(BaseModel):
+    created_at: datetime.datetime = pydantic.Field(title="Created At")
+    updated_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Updated At"
+    )
+    archived_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Archived At"
+    )
+    owner: uuid.UUID = pydantic.Field(title="Owner")
+    user_group: str | None = pydantic.Field(default=None, title="User Group")
+    name: str = pydantic.Field(title="Name")
+    goliat_project_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Goliat Project Id"
+    )
+    pipeline_revision_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="Pipeline Revision Id"
+    )
+    status: AnnotationRunStatus | None = pydantic.Field(default=None, title="Status")
+    dataset_id: uuid.UUID | None = pydantic.Field(default=None, title="Dataset Id")
+    subset_id: uuid.UUID | None = pydantic.Field(default=None, title="Subset Id")
+    annotation_run_type: AnnotationRunType | None = pydantic.Field(
+        default=None, title="Run Type"
+    )
+    ml_annotation_model_id: uuid.UUID | None = pydantic.Field(
+        default=None, title="ML Annotation Model Id"
+    )
+    is_multinode: bool = pydantic.Field(default=False, title="Is Multinode")
+    config: AnnotationRunConfig | None = pydantic.Field(default=None, title="Config")
+    id: uuid.UUID = pydantic.Field(title="Id")
+    annotation_run_url: str | None = pydantic.Field(
+        default=None, title="Annotation Run Url"
+    )
+    completed_at: datetime.datetime | None = pydantic.Field(
+        default=None, title="Completed At"
+    )
+    nodes: list[AnnotationRunNode] | None = pydantic.Field(default=None, title="Nodes")
+    ml_annotation_models: list[MLAnnotationModel] | None = pydantic.Field(
+        default=None, title="ML Annotation Models"
+    )
+
+    model_config = pydantic.ConfigDict(extra="ignore")
 
 
 class BulkAttributeCreate(AttributeCreate):